+++ pkg/scalers/aws_sqs_queue_test.go
@@ -28,6 +28,7 @@ type parseAWSSQSMetadataTestData struct {
 
 type awsSQSMetricIdentifier struct {
 	metadataTestData *parseAWSSQSMetadataTestData
+	scalerIndex      int
 	name             string
 }
 
@@ -131,7 +132,8 @@ var testAWSSQSMetadata = []parseAWSSQSMetadataTestData{
 }
 
 var awsSQSMetricIdentifiers = []awsSQSMetricIdentifier{
-	{&testAWSSQSMetadata[1], "AWS-SQS-Queue-DeleteArtifactQ"},
+	{&testAWSSQSMetadata[1], 0, "s0-AWS-SQS-Queue-DeleteArtifactQ"},
+	{&testAWSSQSMetadata[1], 1, "s1-AWS-SQS-Queue-DeleteArtifactQ"},
 }
 
 func TestSQSParseMetadata(t *testing.T) {
@@ -148,7 +150,7 @@ func TestSQSParseMetadata(t *testing.T) {
 
 func TestAWSSQSGetMetricSpecForScaling(t *testing.T) {
 	for _, testData := range awsSQSMetricIdentifiers {
-		meta, err := parseAwsSqsQueueMetadata(&ScalerConfig{TriggerMetadata: testData.metadataTestData.metadata, ResolvedEnv: testAWSSQSAuthentication, AuthParams: testData.metadataTestData.authParams})
+		meta, err := parseAwsSqsQueueMetadata(&ScalerConfig{TriggerMetadata: testData.metadataTestData.metadata, ResolvedEnv: testAWSSQSAuthentication, AuthParams: testData.metadataTestData.authParams, ScalerIndex: testData.scalerIndex})
 		if err != nil {
 			t.Fatal("Could not parse metadata:", err)
 		}