+++ pkg/scalers/kafka_scaler.go
@@ -43,6 +43,8 @@ type kafkaMetadata struct {
 	cert      string
 	key       string
 	ca        string
+
+	scalerIndex int
 }
 
 type offsetResetPolicy string
@@ -201,7 +203,7 @@ func parseKafkaMetadata(config *ScalerConfig) (kafkaMetadata, error) {
 		}
 		meta.version = version
 	}
-
+	meta.scalerIndex = config.ScalerIndex
 	return meta, nil
 }
 
@@ -351,7 +353,7 @@ func (s *kafkaScaler) GetMetricSpecForScaling() []v2beta2.MetricSpec {
 	targetMetricValue := resource.NewQuantity(s.metadata.lagThreshold, resource.DecimalSI)
 	externalMetric := &v2beta2.ExternalMetricSource{
 		Metric: v2beta2.MetricIdentifier{
-			Name: kedautil.NormalizeString(fmt.Sprintf("%s-%s-%s", "kafka", s.metadata.topic, s.metadata.group)),
+			Name: GenerateMetricNameWithIndex(s.metadata.scalerIndex, kedautil.NormalizeString(fmt.Sprintf("%s-%s-%s", "kafka", s.metadata.topic, s.metadata.group))),
 		},
 		Target: v2beta2.MetricTarget{
 			Type:         v2beta2.AverageValueMetricType,