+++ pkg/scalers/prometheus_scaler_test.go
@@ -13,6 +13,7 @@ type parsePrometheusMetadataTestData struct {
 
 type prometheusMetricIdentifier struct {
 	metadataTestData *parsePrometheusMetadataTestData
+	scalerIndex      int
 	name             string
 }
 
@@ -33,7 +34,8 @@ var testPromMetadata = []parsePrometheusMetadataTestData{
 }
 
 var prometheusMetricIdentifiers = []prometheusMetricIdentifier{
-	{&testPromMetadata[1], "prometheus-http_requests_total"},
+	{&testPromMetadata[1], 0, "s0-prometheus-http_requests_total"},
+	{&testPromMetadata[1], 1, "s1-prometheus-http_requests_total"},
 }
 
 type prometheusAuthMetadataTestData struct {
@@ -79,7 +81,7 @@ func TestPrometheusParseMetadata(t *testing.T) {
 
 func TestPrometheusGetMetricSpecForScaling(t *testing.T) {
 	for _, testData := range prometheusMetricIdentifiers {
-		meta, err := parsePrometheusMetadata(&ScalerConfig{TriggerMetadata: testData.metadataTestData.metadata})
+		meta, err := parsePrometheusMetadata(&ScalerConfig{TriggerMetadata: testData.metadataTestData.metadata, ScalerIndex: testData.scalerIndex})
 		if err != nil {
 			t.Fatal("Could not parse metadata:", err)
 		}