+++ pkg/scalers/redis_scaler_test.go
@@ -21,6 +21,7 @@ type parseRedisMetadataTestData struct {
 
 type redisMetricIdentifier struct {
 	metadataTestData *parseRedisMetadataTestData
+	scalerIndex      int
 	name             string
 }
 
@@ -51,7 +52,8 @@ var testRedisMetadata = []parseRedisMetadataTestData{
 	{map[string]string{"listName": "mylist", "listLength": "0"}, true, map[string]string{"host": "localhost"}}}
 
 var redisMetricIdentifiers = []redisMetricIdentifier{
-	{&testRedisMetadata[1], "redis-mylist"},
+	{&testRedisMetadata[1], 0, "s0-redis-mylist"},
+	{&testRedisMetadata[1], 1, "s1-redis-mylist"},
 }
 
 func TestRedisParseMetadata(t *testing.T) {
@@ -70,7 +72,7 @@ func TestRedisParseMetadata(t *testing.T) {
 
 func TestRedisGetMetricSpecForScaling(t *testing.T) {
 	for _, testData := range redisMetricIdentifiers {
-		meta, err := parseRedisMetadata(&ScalerConfig{TriggerMetadata: testData.metadataTestData.metadata, ResolvedEnv: testRedisResolvedEnv, AuthParams: testData.metadataTestData.authParams}, parseRedisAddress)
+		meta, err := parseRedisMetadata(&ScalerConfig{TriggerMetadata: testData.metadataTestData.metadata, ResolvedEnv: testRedisResolvedEnv, AuthParams: testData.metadataTestData.authParams, ScalerIndex: testData.scalerIndex}, parseRedisAddress)
 		if err != nil {
 			t.Fatal("Could not parse metadata:", err)
 		}