+++ pkg/scalers/stan_scaler.go
@@ -50,6 +50,7 @@ type stanMetadata struct {
 	durableName                  string
 	subject                      string
 	lagThreshold                 int64
+	scalerIndex                  int
 }
 
 const (
@@ -106,6 +107,7 @@ func parseStanMetadata(config *ScalerConfig) (stanMetadata, error) {
 		meta.lagThreshold = t
 	}
 
+	meta.scalerIndex = config.ScalerIndex
 	return meta, nil
 }
 
@@ -196,9 +198,10 @@ func (s *stanScaler) hasPendingMessage() bool {
 
 func (s *stanScaler) GetMetricSpecForScaling() []v2beta2.MetricSpec {
 	targetMetricValue := resource.NewQuantity(s.metadata.lagThreshold, resource.DecimalSI)
+	metricName := kedautil.NormalizeString(fmt.Sprintf("%s-%s-%s-%s", "stan", s.metadata.queueGroup, s.metadata.durableName, s.metadata.subject))
 	externalMetric := &v2beta2.ExternalMetricSource{
 		Metric: v2beta2.MetricIdentifier{
-			Name: kedautil.NormalizeString(fmt.Sprintf("%s-%s-%s-%s", "stan", s.metadata.queueGroup, s.metadata.durableName, s.metadata.subject)),
+			Name: GenerateMetricNameWithIndex(s.metadata.scalerIndex, metricName),
 		},
 		Target: v2beta2.MetricTarget{
 			Type:         v2beta2.AverageValueMetricType,